+++ Server/system.Core/ProjectionModels/Schedule/ScheduleFilter.cs
@@ -2,7 +2,7 @@
 
 public record ScheduleFilter
 (
-    IEnumerable<long> WorkerIds,
+    IEnumerable<int> WorkerIds,
     IEnumerable<long> ShiftIds,
     string? SortBy,
     int Page,
