+++ Server/system.Core/ProjectionModels/WorkInOrder/WorkInOrderFilter.cs
@@ -4,7 +4,7 @@ public record WorkInOrderFilter
 (
     IEnumerable<long> OrderIds,
     IEnumerable<long> JobIds,
-    IEnumerable<long> WorkerIds,
+    IEnumerable<int> WorkerIds,
     IEnumerable<int> StatusIds,
     string? SortBy,
     int Page,
