+++ Server/system.Core/ProjectionModels/WorkProposal/WorkProposalFilter.cs
@@ -4,7 +4,7 @@ public record WorkProposalFilter
 (
     IEnumerable<long> OrderIds,
     IEnumerable<long> JobIds,
-    IEnumerable<long> WorkerIds,
+    IEnumerable<int> WorkerIds,
     IEnumerable<int> StatusIds,
     string? SortBy,
     int Page,
